refactor(gui): extract batch flash and batch authorize into batch.rs
Fifth and final cut at lib.rs. Moves the eight batch commands, the slot worker
extracted two commits ago, BatchFlashState/BatchAuthState/BatchSlot, the Excel
allocator session and its SlotSessionGuard, and the tests covering them.

The allocator locking rule is the reason these belong in one module: a single
mutex guards both acquisition in batch_auth_start and release from the last
SlotSessionGuard drop, with the invariant `alloc.is_some() => active > 0`
meaning the sheet is locked exactly while slots run. The module header now
states that next to the code enforcing it.

lib.rs 1873 -> 964. Over the five refactor commits: 4630 -> 964, with the file
now holding app setup, the flash command, the batch archive, and external-URL
opening. 45 commands registered, 363 workspace tests pass, clippy clean.
